PAP-12424 Work Timeline — share timeline DTOs + add issueTitle to spans
Phase C needs the timeline result types on the client and the truncated task
title in the hover tooltip (board-locked design shows no ID on the bar). Move
the WorkTimeline DTOs (actors/spans/events/edges/result) into @paperclipai/shared
so the aggregation service and the UI consume one contract, and add issueTitle to
each span so the frontend needs no N+1 issue fetch. Additive: existing tests use
objectContaining and still pass.
