+++ packages/shared/src/index.ts
@@ -752,6 +752,14 @@ export type {
   LiveEvent,
   DashboardRunActivityDay,
   DashboardSummary,
+  TimelineActorType,
+  TimelineEventKind,
+  TimelineEdgeKind,
+  WorkTimelineActor,
+  WorkTimelineSpan,
+  WorkTimelineEvent,
+  WorkTimelineEdge,
+  WorkTimelineResult,
   ActivityEvent,
   UserProfileActivitySummary,
   UserProfileAgentUsage,
